feat(frontend): move Login and Signup components to pages directory
